pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "colorformer"
version = "0.1.0"
description = "Modular Colorformer image colorization package"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch",
  "torchvision",
  "matplotlib",
  "kornia",
  "kagglehub",
  "tqdm",
  "Pillow",
  "numpy",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
